+++ packages/platforms/react-native/android/src/main/java/com/bugsnag/android/performance/NativeBugsnagPerformanceImpl.java
@@ -1,17 +1,21 @@
 package com.bugsnag.reactnative.performance;
 
-import android.content.Context;
+import android.annotation.SuppressLint;
 import android.content.pm.PackageInfo;
-import android.content.pm.PackageManager;
 import android.os.Build;
-import androidx.annotation.NonNull;
+import android.os.Handler;
+import android.os.Looper;
+import android.os.SystemClock;
+
 import androidx.annotation.Nullable;
+
 import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
+import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.WritableArray;
 import com.facebook.react.bridge.WritableMap;
-import java.security.SecureRandom;
+
 import java.io.File;
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
@@ -19,17 +23,62 @@ import java.io.OutputStreamWriter;
 import java.io.Writer;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
+import java.security.SecureRandom;
+import java.util.Iterator;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.TimeUnit;
 
+import com.bugsnag.android.performance.BugsnagPerformance;
+import com.bugsnag.android.performance.SpanOptions;
+
+import com.bugsnag.android.performance.internal.BugsnagClock;
+import com.bugsnag.android.performance.internal.EncodingUtils;
+import com.bugsnag.android.performance.internal.SpanFactory;
+import com.bugsnag.android.performance.internal.SpanImpl;
+import com.bugsnag.android.performance.internal.processing.ImmutableConfig;
+
+@SuppressLint("RestrictedApi")
 class NativeBugsnagPerformanceImpl {
-  
+
   static final String MODULE_NAME = "BugsnagReactNativePerformance";
-  
+
   private final ReactApplicationContext reactContext;
 
   private final SecureRandom random = new SecureRandom();
 
+  private boolean isNativePerformanceAvailable = false;
+
+  private boolean isCleanupTaskScheduled = false;
+
+  /**
+   * A map of open native spans, keyed by the span ID and trace ID,
+   * so that they can be retrieved and closed/discarded from JS.
+   */
+  private final ConcurrentHashMap<String, SpanImpl> openSpans = new ConcurrentHashMap<>();
+
+  private final Handler mainHandler = new Handler(Looper.getMainLooper());
+
+  private final Runnable spanCleanupTask = new Runnable() {
+    @Override
+    public void run() {
+      discardLongRunningSpans();
+      // Reschedule the task to run again after one hour
+      mainHandler.postDelayed(this, TimeUnit.HOURS.toMillis(1));
+    }
+  };
+
   public NativeBugsnagPerformanceImpl(ReactApplicationContext reactContext) {
     this.reactContext = reactContext;
+
+    try {
+      BugsnagPerformance.INSTANCE.getInstrumentedAppState$internal().getConfig$internal();
+      isNativePerformanceAvailable = true;
+    }
+    catch (LinkageError e) {
+      // do nothing, Android Performance SDK is not installed or is incompatible
+    }
   }
 
   WritableMap getDeviceInfo() {
@@ -79,6 +128,168 @@ class NativeBugsnagPerformanceImpl {
 
   void requestEntropyAsync(Promise promise) {
     promise.resolve(requestEntropy());
+  }
+
+  boolean isNativePerformanceAvailable() {
+    return isNativePerformanceAvailable;
+  }
+
+  @Nullable
+  WritableMap attachToNativeSDK() {
+    if (!isNativePerformanceAvailable) {
+      return null;
+    }
+
+    ImmutableConfig nativeConfig = BugsnagPerformance.INSTANCE.getInstrumentedAppState$internal().getConfig$internal();
+    if (nativeConfig == null) {
+      return null;
+    }
+
+    if (!isCleanupTaskScheduled) {
+      mainHandler.postDelayed(spanCleanupTask, TimeUnit.HOURS.toMillis(1));
+      isCleanupTaskScheduled = true;
+    }
+    
+    WritableMap result = Arguments.createMap();
+    result.putString("apiKey", nativeConfig.getApiKey());
+    result.putString("endpoint", nativeConfig.getEndpoint());
+    result.putString("releaseStage", nativeConfig.getReleaseStage());
+    result.putString("serviceName", nativeConfig.getServiceName());
+    result.putInt("attributeCountLimit", nativeConfig.getAttributeCountLimit());
+    result.putInt("attributeStringValueLimit", nativeConfig.getAttributeStringValueLimit());
+    result.putInt("attributeArrayLengthLimit", nativeConfig.getAttributeArrayLengthLimit());
+
+    String appVersion = nativeConfig.getAppVersion();
+    if (appVersion != null) {
+      result.putString("appVersion", nativeConfig.getAppVersion());
+    }
+
+    Double samplingProbability = nativeConfig.getSamplingProbability();
+    if (samplingProbability != null) {
+      result.putDouble("samplingProbability", samplingProbability);
+    }
+
+    Set<String> enabledReleaseStages = nativeConfig.getEnabledReleaseStages();
+    if (enabledReleaseStages != null) {
+      result.putArray("enabledReleaseStages", Arguments.fromArray(enabledReleaseStages.toArray(new String[0])));
+    }
+
+    return result;
+  }
+
+  @Nullable
+  WritableMap startNativeSpan(String name, ReadableMap options) {
+    if (!isNativePerformanceAvailable) {
+      return null;
+    }
+
+    SpanOptions spanOptions = readableMapToSpanOptions(options);
+    SpanFactory spanFactory = BugsnagPerformance.INSTANCE.getInstrumentedAppState$internal().getSpanFactory();
+    SpanImpl nativeSpan = spanFactory.createCustomSpan(name, spanOptions);
+
+    // all span attributes are set from JS, with the exception of the bugsnag.sampling.p attribute
+    // this needs to be preserved as it may not be re-populated when the span is processed
+    Iterator<Map.Entry<String, Object>> entries = nativeSpan.getAttributes().getEntries$internal().iterator();
+    while (entries.hasNext()) {
+      Map.Entry<String, Object> entry = entries.next();
+      if (!entry.getKey().equals("bugsnag.sampling.p")) {
+        entries.remove();
+      }
+    }
+
+    String spanKey = EncodingUtils.toHexString(nativeSpan.getSpanId()) + EncodingUtils.toHexString(nativeSpan.getTraceId());
+    openSpans.put(spanKey, nativeSpan);
+
+    return nativeSpanToJsSpan(nativeSpan);
+  }
+
+  void markNativeSpanEndTime(String spanId, String traceId, double endTime) {
+    SpanImpl nativeSpan = openSpans.get(spanId + traceId);
+    if (nativeSpan != null) {
+      long nativeEndTime = BugsnagClock.INSTANCE.unixNanoTimeToElapsedRealtime((long)endTime);
+      nativeSpan.markEndTime$internal(nativeEndTime);
+    }
+  }
+
+  void endNativeSpan(String spanId, String traceId, double endTime, ReadableMap jsAttributes, Promise promise) {
+    SpanImpl nativeSpan = openSpans.remove(spanId + traceId);
+    if (nativeSpan == null) {
+      promise.resolve(null);
+      return;
+    }
+
+    ReactNativeSpanAttributes.setAttributesFromReadableMap(nativeSpan.getAttributes(), jsAttributes);
+    long nativeEndTime = BugsnagClock.INSTANCE.unixNanoTimeToElapsedRealtime((long)endTime);
+    if (nativeEndTime > nativeSpan.getEndTime$internal()) {
+      nativeSpan.markEndTime$internal(nativeEndTime);
+    }
+
+    nativeSpan.sendForProcessing$internal();
+    promise.resolve(null);
+  }
+
+  void discardNativeSpan(String spanId, String traceId, Promise promise) {
+    SpanImpl nativeSpan = openSpans.remove(spanId + traceId);
+    if (nativeSpan != null) {
+      nativeSpan.discard();
+    }
+
+    promise.resolve(null);
+  }
+
+  private WritableMap nativeSpanToJsSpan(SpanImpl nativeSpan) {
+    WritableMap span = Arguments.createMap();
+    span.putString("name", nativeSpan.getName());
+    span.putString("id", EncodingUtils.toHexString(nativeSpan.getSpanId()));
+    span.putString("traceId", EncodingUtils.toHexString(nativeSpan.getTraceId()));
+
+    long unixNanoStartTime = BugsnagClock.INSTANCE.elapsedNanosToUnixTime(nativeSpan.getStartTime$internal());
+    span.putDouble("startTime", (double)unixNanoStartTime);
+
+    long parentSpanId = nativeSpan.getParentSpanId();
+    if (parentSpanId != 0L) {
+      span.putString("parentSpanId", EncodingUtils.toHexString(parentSpanId));
+    }
+
+    return span;
+  }
+
+  private SpanOptions readableMapToSpanOptions(ReadableMap jsOptions) {
+    SpanOptions spanOptions = SpanOptions.DEFAULTS
+      .setFirstClass(true)
+      .makeCurrentContext(false)
+      .within(null);
+
+    if (jsOptions.hasKey("startTime")) {
+      double startTime = jsOptions.getDouble("startTime");
+      long nativeStartTime = BugsnagClock.INSTANCE.unixNanoTimeToElapsedRealtime((long)startTime);
+      spanOptions = spanOptions.startTime(nativeStartTime);
+    }
+
+    ReadableMap parentContext;
+    if (jsOptions.hasKey("parentContext") && (parentContext = jsOptions.getMap("parentContext")) != null) {
+      ReactNativeSpanContext nativeParentContext = new ReactNativeSpanContext(
+        parentContext.getString("id"),
+        parentContext.getString("traceId")
+      );
+
+      spanOptions = spanOptions.within(nativeParentContext);
+    }
+
+    return spanOptions;
+  }
+
+  private void discardLongRunningSpans() {
+    long oneHourAgo = SystemClock.elapsedRealtimeNanos() - TimeUnit.HOURS.toNanos(1);
+    Iterator<Map.Entry<String, SpanImpl>> entries = openSpans.entrySet().iterator();
+    while (entries.hasNext()) {
+      Map.Entry<String, SpanImpl> entry = entries.next();
+      SpanImpl span = entry.getValue();
+      if (span.getStartTime$internal() < oneHourAgo) {
+        entries.remove();
+        span.discard();
+      }
+    }
   }
 
   WritableMap getNativeConstants() {
@@ -171,7 +382,7 @@ class NativeBugsnagPerformanceImpl {
     }
   }
 
-  void writeFile(String path, String data, String encoding, Promise promise){
+  void writeFile(String path, String data, String encoding, Promise promise) {
     try(
       FileOutputStream fout = new FileOutputStream(path);
       Writer w = new OutputStreamWriter(fout, encoding);