+++ packages/platforms/react-native/lib/auto-instrumentation/app-start-plugin.tsx
@@ -11,6 +11,7 @@ import React from 'react'
 import type { AppRegistry, WrapperComponentProvider } from 'react-native'
 import type { ReactNativeConfiguration } from '../config'
 import { createAppStartSpan } from '../create-app-start-span'
+import type { ReactNativeSpanFactory } from '../span-factory'
 
 interface WrapperProps {
   children: ReactNode
@@ -21,7 +22,7 @@ export const isWrapperComponentProvider = (value: unknown): value is WrapperComp
 
 export class AppStartPlugin implements Plugin<ReactNativeConfiguration> {
   private readonly appStartTime: number
-  private readonly spanFactory: SpanFactory<ReactNativeConfiguration>
+  private readonly spanFactory: ReactNativeSpanFactory
   private readonly clock: Clock
   private readonly appRegistry: typeof AppRegistry
   private readonly setAppState: SetAppState
@@ -29,7 +30,7 @@ export class AppStartPlugin implements Plugin<ReactNativeConfiguration> {
 
   constructor (
     appStartTime: number,
-    spanFactory: SpanFactory<ReactNativeConfiguration>,
+    spanFactory: ReactNativeSpanFactory,
     clock: Clock,
     appRegistry: typeof AppRegistry,
     setAppState: SetAppState,