+++ packages/platforms/react-native/tests/native.test.ts
@@ -11,43 +11,28 @@ jest.mock('react-native', () => {
   }
 })
 
-describe('React Native turbomodule is null so implementation falls back to stub functions', () => {
+describe('NativeBugsnagPerformance', () => {
   afterAll(() => {
     jest.resetModules()
   })
-  it('getDeviceInfo returns undefined', () => {
+
+  it('falls back to stub functions when native module is null', async () => {
     expect(NativeBugsnagPerformance.getDeviceInfo()).toBeUndefined()
-  })
-
-  it('requestEntropy returns an empty string', () => {
     expect(NativeBugsnagPerformance.requestEntropy()).toBe('')
-  })
-
-  it('requestEntropyAsync returns an empty string', async () => {
     expect(await NativeBugsnagPerformance.requestEntropyAsync()).toBe('')
-  })
-
-  it('getNativeConstants returns an empty string', () => {
     expect(NativeBugsnagPerformance.getNativeConstants()).toStrictEqual({ CacheDir: '', DocumentDir: '' })
-  })
-
-  it('exists returns false', async () => {
     expect(await NativeBugsnagPerformance.exists('')).toBe(false)
-  })
-
-  it('isDir returns false', async () => {
     expect(await NativeBugsnagPerformance.isDir('')).toBe(false)
-  })
-
-  it('ls returns an empty array', async () => {
     expect(await NativeBugsnagPerformance.ls('')).toStrictEqual([])
-  })
-
-  it('mkdir returns an empty string', async () => {
     expect(await NativeBugsnagPerformance.mkdir('')).toBe('')
-  })
-
-  it('readFile returns an empty string', async () => {
     expect(await NativeBugsnagPerformance.readFile('', '')).toBe('')
+    await expect(NativeBugsnagPerformance.unlink('')).resolves.toBeUndefined()
+    await expect(NativeBugsnagPerformance.writeFile('', '', '')).resolves.toBeUndefined()
+    expect(NativeBugsnagPerformance.isNativePerformanceAvailable()).toBe(false)
+    expect(NativeBugsnagPerformance.attachToNativeSDK()).toBeNull()
+    expect(NativeBugsnagPerformance.startNativeSpan('', {})).toStrictEqual({ name: '', id: '', traceId: '', startTime: 0, parentSpanId: '' })
+    await expect(NativeBugsnagPerformance.endNativeSpan('', '', 0, {})).resolves.toBeUndefined()
+    await expect(NativeBugsnagPerformance.discardNativeSpan('', '')).resolves.toBeUndefined()
+    expect(() => { NativeBugsnagPerformance.markNativeSpanEndTime('', '', 0) }).not.toThrow()
   })
 })